+++ datacontract/data_contract.py
@@ -27,6 +27,7 @@ class DataContract:
         data_contract: OpenDataContractStandard = None,
         schema_location: str = None,
         server: str = None,
+        schema_name: str = "all",
         publish_url: str = None,
         spark: "SparkSession" = None,
         duckdb_connection: "DuckDBPyConnection" = None,
@@ -40,6 +41,7 @@ class DataContract:
         self._data_contract = data_contract
         self._schema_location = schema_location
         self._server = server
+        self._schema_name = schema_name
         self._publish_url = publish_url
         self._publish_test_results = publish_test_results
         self._spark = spark
@@ -120,7 +122,9 @@ class DataContract:
                 inline_definitions=self._inline_definitions,
             )
 
-            execute_data_contract_test(data_contract, run, self._server, self._spark, self._duckdb_connection)
+            execute_data_contract_test(
+                data_contract, run, self._server, self._spark, self._duckdb_connection, schema_name=self._schema_name
+            )
 
         except DataContractException as e:
             run.checks.append(